+++ packages/expo-video/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Add `seekTo` method on Android and iOS. ([#28522](https://github.com/expo/expo/pull/28522) by [@justjoostnl](https://github.com/justjoostnl))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others